refactor(battle): extract end-of-turn pipeline helper (#991)
