+++ packages/battle/src/engine/BattleEngine.ts
@@ -40,6 +40,7 @@ import {
   createPokemonSnapshot,
   getPokemonName,
 } from "../utils";
+import { processEndOfTurnPipeline } from "./BattleEndOfTurnPipeline";
 
 /**
  * Moves that can be used while the user is asleep.
@@ -4310,773 +4311,40 @@ export class BattleEngine implements BattleEventEmitter {
   }
 
   private processEndOfTurn(): void {
-    const effectOrder = this.ruleset.getEndOfTurnOrder();
-
-    // Bug #484 fix: Multiple EoT cases (weather-healing, shed-skin, speed-boost, etc.) all
-    // call applyAbility("on-turn-end") for every active Pokemon. The gen ruleset's ability
-    // handler fires whatever EoT ability the Pokemon has (e.g., Speed Boost) regardless of
-    // which EoT case triggered it. This Set ensures each Pokemon's on-turn-end ability fires
-    // at most once per turn, no matter how many EoT ability-dispatching cases appear.
-    // Source: pret/pokeemerald src/battle_util.c — ABILITYEFFECT_ENDTURN fires once per Pokemon
-    const abilityEndOfTurnFired = new Set<string>();
-
-    for (const effect of effectOrder) {
-      switch (effect) {
-        case "weather-damage":
-          this.processWeatherDamage();
-          break;
-        case "weather-countdown":
-          this.processWeatherCountdown();
-          break;
-        case "terrain-countdown":
-          this.processTerrainCountdown();
-          break;
-        case "status-damage":
-          this.processStatusDamage();
-          break;
-        case "screen-countdown":
-          this.processScreenCountdown();
-          break;
-        case "tailwind-countdown":
-          this.processTailwindCountdown();
-          break;
-        case "trick-room-countdown":
-          this.processTrickRoomCountdown();
-          break;
-        case "leftovers":
-          this.processHeldItemEndOfTurn();
-          break;
-        case "leech-seed":
-          this.processLeechSeed();
-          break;
-        case "perish-song":
-          this.processPerishSong();
-          break;
-        case "curse":
-          this.processCurse();
-          break;
-        case "nightmare":
-          this.processNightmare();
-          break;
-        case "bind":
-          this.processBindDamage();
-          break;
-        case "salt-cure":
-          this.processSaltCureEoT();
-          break;
-        case "defrost":
-          this.processDefrost();
-          break;
-        case "safeguard-countdown":
-          this.processSafeguardCountdown();
-          break;
-        case "mystery-berry":
-          this.processMysteryBerry();
-          break;
-        case "stat-boosting-items":
-          this.processStatBoostingItems();
-          break;
-        case "healing-items":
-          this.processHealingItems();
-          break;
-        case "encore-countdown":
-          this.processEncoreCountdown();
-          break;
-        case "weather-healing": {
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            const pokeKey = `${side.index}-0`;
-            if (abilityEndOfTurnFired.has(pokeKey)) continue;
-            abilityEndOfTurnFired.add(pokeKey);
-            const opponent = this.getOpponentActive(side.index);
-            const result = this.ruleset.applyAbility("on-turn-end", {
-              pokemon: active,
-              opponent: opponent ?? undefined,
-              state: this.state,
-              rng: this.state.rng,
-              trigger: "on-turn-end",
-            });
-            if (result.activated) {
-              this.processAbilityResult(result, active, opponent ?? active, side.index);
-            }
-          }
-          break;
-        }
-        case "shed-skin": {
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            const pokeKey = `${side.index}-0`;
-            if (abilityEndOfTurnFired.has(pokeKey)) continue;
-            abilityEndOfTurnFired.add(pokeKey);
-            const opponent = this.getOpponentActive(side.index);
-            const result = this.ruleset.applyAbility("on-turn-end", {
-              pokemon: active,
-              opponent: opponent ?? undefined,
-              state: this.state,
-              rng: this.state.rng,
-              trigger: "on-turn-end",
-            });
-            if (result.activated) {
-              this.processAbilityResult(result, active, opponent ?? active, side.index);
-            }
-          }
-          break;
-        }
-        case "poison-heal": {
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            const pokeKey = `${side.index}-0`;
-            if (abilityEndOfTurnFired.has(pokeKey)) continue;
-            abilityEndOfTurnFired.add(pokeKey);
-            const opponent = this.getOpponentActive(side.index);
-            const result = this.ruleset.applyAbility("on-turn-end", {
-              pokemon: active,
-              opponent: opponent ?? undefined,
-              state: this.state,
-              rng: this.state.rng,
-              trigger: "on-turn-end",
-            });
-            if (result.activated) {
-              this.processAbilityResult(result, active, opponent ?? active, side.index);
-            }
-          }
-          break;
-        }
-        case "bad-dreams": {
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            const pokeKey = `${side.index}-0`;
-            if (abilityEndOfTurnFired.has(pokeKey)) continue;
-            abilityEndOfTurnFired.add(pokeKey);
-            const opponent = this.getOpponentActive(side.index);
-            const result = this.ruleset.applyAbility("on-turn-end", {
-              pokemon: active,
-              opponent: opponent ?? undefined,
-              state: this.state,
-              rng: this.state.rng,
-              trigger: "on-turn-end",
-            });
-            if (result.activated) {
-              this.processAbilityResult(result, active, opponent ?? active, side.index);
-            }
-          }
-          break;
-        }
-        case "speed-boost": {
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            const pokeKey = `${side.index}-0`;
-            if (abilityEndOfTurnFired.has(pokeKey)) continue;
-            abilityEndOfTurnFired.add(pokeKey);
-            const opponent = this.getOpponentActive(side.index);
-            const result = this.ruleset.applyAbility("on-turn-end", {
-              pokemon: active,
-              opponent: opponent ?? undefined,
-              state: this.state,
-              rng: this.state.rng,
-              trigger: "on-turn-end",
-            });
-            if (result.activated) {
-              this.processAbilityResult(result, active, opponent ?? active, side.index);
-            }
-          }
-          break;
-        }
-        case "slow-start-countdown": {
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            // Slow Start: decrement turnsLeft on the slow-start volatile each EoT.
-            // No ability check: the volatile should tick down even if the ability was
-            // temporarily changed (e.g., by Skill Swap). The stat-halving in damage/speed
-            // calcs already requires both the ability AND the volatile to be present.
-            // Source: Pokemon Showdown Gen 4 mod — Slow Start countdown ticks volatile
-            // When turnsLeft reaches 0, remove the volatile so the Attack/Speed halving stops.
-            // Source: Pokemon Showdown Gen 4 mod — Slow Start countdown
-            // Source: Bulbapedia — Slow Start: halves Attack and Speed for 5 turns
-            const slowStart = active.volatileStatuses.get("slow-start");
-            if (slowStart && slowStart.turnsLeft > 0) {
-              slowStart.turnsLeft -= 1;
-              if (slowStart.turnsLeft === 0) {
-                active.volatileStatuses.delete("slow-start");
-                const pokeName = active.pokemon.nickname ?? String(active.pokemon.speciesId);
-                this.emit({
-                  type: "volatile-end",
-                  side: side.index,
-                  pokemon: getPokemonName(active),
-                  volatile: "slow-start",
-                });
-                this.emit({
-                  type: "message",
-                  text: `${pokeName}'s Slow Start wore off!`,
-                });
-              }
-            }
-          }
-          break;
-        }
-        case "toxic-orb-activation":
-        case "flame-orb-activation": {
-          // Only trigger the specific orb item, NOT all end-of-turn items.
-          // Fix for bug #600: previously called applyHeldItem("end-of-turn") unfiltered,
-          // causing Leftovers/Black Sludge to activate again during the orb phase.
-          // Source: Showdown data/items.ts — Toxic Orb and Flame Orb are independent
-          //   end-of-turn triggers that only affect the holder of that specific orb.
-          const orbItem = effect === "toxic-orb-activation" ? "toxic-orb" : "flame-orb";
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            if (active.pokemon.heldItem !== orbItem) continue;
-            const itemResult = this.ruleset.applyHeldItem("end-of-turn", {
-              pokemon: active,
-              state: this.state,
-              rng: this.state.rng,
-            });
-            if (itemResult.activated) {
-              this.processItemResult(itemResult, active, side.index);
-            }
-          }
-          break;
-        }
-        case "black-sludge": {
-          // Handled via applyHeldItem("end-of-turn") — same as leftovers, gen ruleset distinguishes
-          if (!this.ruleset.hasHeldItems()) break;
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            if (active.pokemon.heldItem !== "black-sludge") continue;
-            const itemResult = this.ruleset.applyHeldItem("end-of-turn", {
-              pokemon: active,
-              state: this.state,
-              rng: this.state.rng,
-            });
-            if (itemResult.activated) {
-              this.processItemResult(itemResult, active, side.index);
-            }
-          }
-          break;
-        }
-        case "aqua-ring": {
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            if (!active.volatileStatuses.has("aqua-ring")) continue;
-            const maxHp = active.pokemon.calculatedStats?.hp ?? active.pokemon.currentHp;
-            const healAmount = Math.max(1, Math.floor(maxHp / 16));
-            const oldHp = active.pokemon.currentHp;
-            active.pokemon.currentHp = Math.min(maxHp, oldHp + healAmount);
-            const healed = active.pokemon.currentHp - oldHp;
-            if (healed > 0) {
-              this.emit({
-                type: "heal",
-                side: side.index,
-                pokemon: getPokemonName(active),
-                amount: healed,
-                currentHp: active.pokemon.currentHp,
-                maxHp,
-                source: "aqua-ring",
-              });
-            }
-          }
-          break;
-        }
-        case "ingrain": {
-          // Source: Bulbapedia — Ingrain heals 1/16 max HP per turn
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            if (!active.volatileStatuses.has("ingrain")) continue;
-            const maxHp = active.pokemon.calculatedStats?.hp ?? active.pokemon.currentHp;
-            const healAmount = Math.max(1, Math.floor(maxHp / 16));
-            const oldHp = active.pokemon.currentHp;
-            active.pokemon.currentHp = Math.min(maxHp, oldHp + healAmount);
-            const healed = active.pokemon.currentHp - oldHp;
-            if (healed > 0) {
-              this.emit({
-                type: "heal",
-                side: side.index,
-                pokemon: getPokemonName(active),
-                amount: healed,
-                currentHp: active.pokemon.currentHp,
-                maxHp,
-                source: "ingrain",
-              });
-            }
-          }
-          break;
-        }
-        case "wish": {
-          for (const side of this.state.sides) {
-            if (!side.wish?.active) continue;
-            side.wish.turnsLeft--;
-            if (side.wish.turnsLeft <= 0) {
-              const active = side.active[0];
-              if (active && active.pokemon.currentHp > 0) {
-                const maxHp = active.pokemon.calculatedStats?.hp ?? active.pokemon.currentHp;
-                const healAmount = Math.min(side.wish.healAmount, maxHp - active.pokemon.currentHp);
-                if (healAmount > 0) {
-                  active.pokemon.currentHp += healAmount;
-                  this.emit({
-                    type: "heal",
-                    side: side.index,
-                    pokemon: getPokemonName(active),
-                    amount: healAmount,
-                    currentHp: active.pokemon.currentHp,
-                    maxHp,
-                    source: "wish",
-                  });
-                }
-              }
-              side.wish = null;
-            }
-          }
-          break;
-        }
-        case "future-attack": {
-          for (const side of this.state.sides) {
-            if (!side.futureAttack) continue;
-            side.futureAttack.turnsLeft--;
-            if (side.futureAttack.turnsLeft <= 0) {
-              const active = side.active[0];
-              if (active && active.pokemon.currentHp > 0) {
-                let futureDamage = side.futureAttack.damage;
-
-                // Protocol: Gen 2-4 rulesets store pre-calculated damage at use time (non-zero).
-                // Gen 5+ rulesets signal hit-time recalculation by returning true from
-                // recalculatesFutureAttackDamage(), OR by storing 0 as a sentinel.
-                // Source: Bulbapedia — "From Generation V onwards, damage is calculated when
-                //   Future Sight or Doom Desire hits, not when it is used."
-                // Source: Showdown sim/battle-actions.ts — Gen 5+ recalculates future attack damage
-                if (futureDamage === 0 || this.ruleset.recalculatesFutureAttackDamage?.()) {
-                  const sourceSideState = this.state.sides[side.futureAttack.sourceSide];
-                  const sourceActive = sourceSideState.active[0];
-                  if (sourceActive && sourceActive.pokemon.currentHp > 0) {
-                    let moveData: MoveData | undefined;
-                    try {
-                      moveData = this.dataManager.getMove(side.futureAttack.moveId);
-                    } catch {
-                      this.emit({
-                        type: "engine-warning",
-                        message:
-                          `Future attack move "${side.futureAttack.moveId}" data missing while resolving. ` +
-                          "Using stored fallback damage.",
-                      });
-                    }
-                    if (moveData) {
-                      const result = this.ruleset.calculateDamage({
-                        attacker: sourceActive,
-                        defender: active,
-                        move: moveData,
-                        state: this.state,
-                        rng: this.state.rng,
-                        isCrit: false,
-                      });
-                      futureDamage = result.damage;
-                    }
-                  }
-                }
-
-                const clampedDamage = Math.min(futureDamage, active.pokemon.currentHp);
-                active.pokemon.currentHp -= clampedDamage;
-                const maxHp =
-                  active.pokemon.calculatedStats?.hp ?? active.pokemon.currentHp + clampedDamage;
-                this.emit({
-                  type: "damage",
-                  side: side.index,
-                  pokemon: getPokemonName(active),
-                  amount: clampedDamage,
-                  currentHp: active.pokemon.currentHp,
-                  maxHp,
-                  source: side.futureAttack.moveId,
-                });
-              }
-              side.futureAttack = null;
-            }
-          }
-          break;
-        }
-        case "taunt-countdown": {
-          // Taunt volatile countdown — remove when turnsLeft reaches 0
-          // Source: Bulbapedia — "Taunt lasts for 3 turns in Gen 4"
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            const tauntState = active.volatileStatuses.get("taunt");
-            if (!tauntState) continue;
-            if (tauntState.turnsLeft > 0) {
-              tauntState.turnsLeft--;
-              if (tauntState.turnsLeft <= 0) {
-                active.volatileStatuses.delete("taunt");
-                this.emit({
-                  type: "volatile-end",
-                  side: side.index,
-                  pokemon: getPokemonName(active),
-                  volatile: "taunt",
-                });
-              }
-            }
-          }
-          break;
-        }
-        case "disable-countdown": {
-          // Disable volatile countdown — remove when turnsLeft reaches 0
-          // Source: Bulbapedia — "Disable lasts for 4-7 turns in Gen 4"
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            const disableState = active.volatileStatuses.get("disable");
-            if (!disableState) continue;
-            if (disableState.turnsLeft > 0) {
-              disableState.turnsLeft--;
-              if (disableState.turnsLeft <= 0) {
-                active.volatileStatuses.delete("disable");
-                this.emit({
-                  type: "volatile-end",
-                  side: side.index,
-                  pokemon: getPokemonName(active),
-                  volatile: "disable",
-                });
-              }
-            }
-          }
-          break;
-        }
-        case "gravity-countdown": {
-          // Gravity field countdown — deactivate when turnsLeft reaches 0
-          // Source: Showdown Gen 4 mod — Gravity lasts 5 turns
-          if (this.state.gravity.active) {
-            this.state.gravity.turnsLeft--;
-            if (this.state.gravity.turnsLeft <= 0) {
-              this.state.gravity.active = false;
-              this.emit({ type: "message", text: "Gravity returned to normal!" });
-            }
-          }
-          break;
-        }
-        case "magic-room-countdown": {
-          // Magic Room field countdown — deactivate when turnsLeft reaches 0
-          // Source: Showdown magicroom condition — duration: 5
-          if (this.state.magicRoom.active) {
-            this.state.magicRoom.turnsLeft--;
-            if (this.state.magicRoom.turnsLeft <= 0) {
-              this.state.magicRoom.active = false;
-              this.emit({ type: "message", text: "The area returned to normal!" });
-            }
-          }
-          break;
-        }
-        case "wonder-room-countdown": {
-          // Wonder Room field countdown — deactivate when turnsLeft reaches 0
-          // Source: Showdown wonderroom condition — duration: 5
-          if (this.state.wonderRoom.active) {
-            this.state.wonderRoom.turnsLeft--;
-            if (this.state.wonderRoom.turnsLeft <= 0) {
-              this.state.wonderRoom.active = false;
-              this.emit({
-                type: "message",
-                text: "Wonder Room wore off, and Defense and Sp. Def stats returned to normal!",
-              });
-            }
-          }
-          break;
-        }
-        case "yawn-countdown": {
-          // Yawn volatile countdown — inflict sleep when turnsLeft reaches 0
-          // Source: Bulbapedia — Yawn: "causes drowsiness; the target falls asleep at
-          //   the end of the next turn"
-          // Source: Showdown Gen 4 mod — Yawn sets a 1-turn drowsy volatile
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            const yawnState = active.volatileStatuses.get("yawn");
-            if (!yawnState) continue;
-            if (yawnState.turnsLeft > 0) {
-              yawnState.turnsLeft--;
-            }
-            if (yawnState.turnsLeft <= 0) {
-              active.volatileStatuses.delete("yawn");
-              this.emit({
-                type: "volatile-end",
-                side: side.index,
-                pokemon: getPokemonName(active),
-                volatile: "yawn",
-              });
-              // Inflict sleep if the target has no primary status
-              if (active.pokemon.status === null) {
-                this.applyPrimaryStatus(active, "sleep", side.index);
-              }
-            }
-          }
-          break;
-        }
-        case "heal-block-countdown": {
-          // Heal Block volatile countdown — remove when turnsLeft reaches 0
-          // Source: Bulbapedia — Heal Block prevents HP recovery for 5 turns
-          // Source: Showdown Gen 4 mod — Heal Block lasts 5 turns
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            const hbState = active.volatileStatuses.get("heal-block");
-            if (!hbState) continue;
-            if (hbState.turnsLeft > 0) {
-              hbState.turnsLeft--;
-              if (hbState.turnsLeft <= 0) {
-                active.volatileStatuses.delete("heal-block");
-                this.emit({
-                  type: "volatile-end",
-                  side: side.index,
-                  pokemon: getPokemonName(active),
-                  volatile: "heal-block",
-                });
-              }
-            }
-          }
-          break;
-        }
-        case "embargo-countdown": {
-          // Embargo volatile countdown — remove when turnsLeft reaches 0
-          // Source: Bulbapedia — Embargo prevents use of held items for 5 turns
-          // Source: Showdown Gen 4 mod — Embargo lasts 5 turns
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            const embargoState = active.volatileStatuses.get("embargo");
-            if (!embargoState) continue;
-            if (embargoState.turnsLeft > 0) {
-              embargoState.turnsLeft--;
-              if (embargoState.turnsLeft <= 0) {
-                active.volatileStatuses.delete("embargo");
-                this.emit({
-                  type: "volatile-end",
-                  side: side.index,
-                  pokemon: getPokemonName(active),
-                  volatile: "embargo",
-                });
-              }
-            }
-          }
-          break;
-        }
-        case "magnet-rise-countdown": {
-          // Magnet Rise volatile countdown — remove when turnsLeft reaches 0
-          // Source: Bulbapedia — Magnet Rise: "The user levitates for five turns."
-          // Source: Showdown Gen 4 mod — Magnet Rise lasts 5 turns
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            const mrState = active.volatileStatuses.get("magnet-rise");
-            if (!mrState) continue;
-            if (mrState.turnsLeft > 0) {
-              mrState.turnsLeft--;
-              if (mrState.turnsLeft <= 0) {
-                active.volatileStatuses.delete("magnet-rise");
-                this.emit({
-                  type: "volatile-end",
-                  side: side.index,
-                  pokemon: getPokemonName(active),
-                  volatile: "magnet-rise",
-                });
-              }
-            }
-          }
-          break;
-        }
-        case "moody": {
-          // Gen 5+ ability: raises one stat by 2 and lowers another by 1 at EoT.
-          // Source: Showdown sim/abilities.ts — Moody triggers at residual phase
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            const pokeKey = `${side.index}-0`;
-            if (abilityEndOfTurnFired.has(pokeKey)) continue;
-            abilityEndOfTurnFired.add(pokeKey);
-            const opponent = this.getOpponentActive(side.index);
-            const result = this.ruleset.applyAbility("on-turn-end", {
-              pokemon: active,
-              opponent: opponent ?? undefined,
-              state: this.state,
-              rng: this.state.rng,
-              trigger: "on-turn-end",
-            });
-            if (result.activated) {
-              this.processAbilityResult(result, active, opponent ?? active, side.index);
-            }
-          }
-          break;
-        }
-        case "harvest": {
-          // Gen 5+ ability: 50% chance to restore a consumed Berry at EoT (100% in sun).
-          // Source: Showdown sim/abilities.ts — Harvest triggers at residual phase
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            const pokeKey = `${side.index}-0`;
-            if (abilityEndOfTurnFired.has(pokeKey)) continue;
-            abilityEndOfTurnFired.add(pokeKey);
-            const opponent = this.getOpponentActive(side.index);
-            const result = this.ruleset.applyAbility("on-turn-end", {
-              pokemon: active,
-              opponent: opponent ?? undefined,
-              state: this.state,
-              rng: this.state.rng,
-              trigger: "on-turn-end",
-            });
-            if (result.activated) {
-              this.processAbilityResult(result, active, opponent ?? active, side.index);
-            }
-          }
-          break;
-        }
-        case "pickup": {
-          // Gen 5+ ability: may pick up an item used during the turn.
-          // Stub: delegates to ruleset ability hook. No gen currently returns this effect.
-          // Source: Showdown sim/abilities.ts — Pickup triggers at residual phase
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-            const pokeKey = `${side.index}-0`;
-            if (abilityEndOfTurnFired.has(pokeKey)) continue;
-            abilityEndOfTurnFired.add(pokeKey);
-            const opponent = this.getOpponentActive(side.index);
-            const result = this.ruleset.applyAbility("on-turn-end", {
-              pokemon: active,
-              opponent: opponent ?? undefined,
-              state: this.state,
-              rng: this.state.rng,
-              trigger: "on-turn-end",
-            });
-            if (result.activated) {
-              this.processAbilityResult(result, active, opponent ?? active, side.index);
-            }
-          }
-          break;
-        }
-        case "uproar": {
-          // Source: pret/pokeemerald -- Uproar: countdown duration, wake sleeping Pokemon
-          // Source: Bulbapedia — Uproar prevents sleep while the user is in uproar
-          // Note: "uproar" is added to VolatileStatus in core/entities/status.ts
-          //
-          // Bug #494 fix: first decrement all uproar counters, THEN check if any Pokemon
-          // still has the uproar volatile. Only wake sleepers if uproar is still active.
-          // Previously, the wake check ran inside the same loop as the decrement, so
-          // sleepers were woken even when the uproar expired on that turn.
-          const uproarVolatile = "uproar" as import("@pokemon-lib-ts/core").VolatileStatus;
-
-          // Step 1: Decrement uproar counters for all active Pokemon
-          for (const side of this.state.sides) {
-            const active = side.active[0];
-            if (!active || active.pokemon.currentHp <= 0) continue;
-
-            const uproarData = active.volatileStatuses.get(uproarVolatile);
-            if (uproarData) {
-              if (uproarData.turnsLeft !== undefined && uproarData.turnsLeft > 0) {
-                uproarData.turnsLeft--;
-                if (uproarData.turnsLeft === 0) {
-                  active.volatileStatuses.delete(uproarVolatile);
-                  this.emit({
-                    type: "volatile-end",
-                    side: side.index,
-                    pokemon: getPokemonName(active),
-                    volatile: uproarVolatile,
-                  });
-                  this.emit({
-                    type: "message",
-                    text: `${getPokemonName(active)}'s uproar ended!`,
-                  });
-                }
-              }
-            }
-          }
-
-          // Step 2: Check if ANY active Pokemon on either side still has the uproar volatile
-          const anyUproarActive = this.state.sides.some((side) => {
-            const active = side.active[0];
-            return (
-              active && active.pokemon.currentHp > 0 && active.volatileStatuses.has(uproarVolatile)
-            );
-          });
-
-          // Step 3: Only wake sleeping Pokemon if uproar is still ongoing
-          if (anyUproarActive) {
-            for (const side of this.state.sides) {
-              const active = side.active[0];
-              if (!active || active.pokemon.currentHp <= 0) continue;
-              if (active.pokemon.status === "sleep") {
-                // Soundproof blocks Uproar wake-up (Uproar is a sound-based move/effect)
-                // Source: Bulbapedia — Soundproof protects from sound-based effects including Uproar
-                // Source: Showdown sim/battle-actions.ts — Soundproof immunity to Uproar
-                if (this.ruleset.hasAbilities() && active.ability === "soundproof") {
-                  continue;
-                }
-                active.pokemon.status = null;
-                this.emit({
-                  type: "status-cure",
-                  side: side.index,
-                  pokemon: getPokemonName(active),
-                  status: "sleep",
-                });
-                this.emit({
-                  type: "message",
-                  text: `${getPokemonName(active)} woke up due to the uproar!`,
-                });
-              }
-            }
-          }
-          break;
-        }
-        case "grassy-terrain-heal": {
-          // Gen 6+ terrain: heals grounded Pokemon for 1/16 max HP at EoT.
-          // Source: Showdown sim/field.ts — Grassy Terrain heals at residual phase
-          if (this.state.terrain?.type === "grassy") {
-            const terrainResults = this.ruleset.applyTerrainEffects(this.state);
-            for (const result of terrainResults) {
-              const active = this.getActiveMutable(result.side);
-              if (active && active.pokemon.currentHp > 0) {
-                const healAmount = result.healAmount ?? 0;
-                if (healAmount > 0) {
-                  const maxHp = active.pokemon.calculatedStats?.hp ?? active.pokemon.currentHp;
-                  const healed = Math.min(healAmount, maxHp - active.pokemon.currentHp);
-                  if (healed > 0) {
-                    active.pokemon.currentHp += healed;
-                    this.emit({
-                      type: "heal",
-                      side: result.side,
-                      pokemon: result.pokemon,
-                      amount: healed,
-                      currentHp: active.pokemon.currentHp,
-                      maxHp,
-                      source: "grassy-terrain",
-                    });
-                  }
-                }
-                if (result.message) {
-                  this.emit({ type: "message", text: result.message });
-                }
-              }
-            }
-          }
-          break;
-        }
-        default:
-          // Remaining effects not yet implemented
-          break;
-      }
-
-      this.checkMidTurnFaints();
-      if (this.state.ended) return;
-    }
-
-    // Increment turnsOnField for all active pokemon
-    for (const side of this.state.sides) {
-      for (const active of side.active) {
-        if (active) active.turnsOnField++;
-      }
-    }
+    processEndOfTurnPipeline({
+      state: this.state,
+      ruleset: this.ruleset,
+      dataManager: this.dataManager,
+      emit: (event) => this.emit(event),
+      processWeatherDamage: () => this.processWeatherDamage(),
+      processWeatherCountdown: () => this.processWeatherCountdown(),
+      processTerrainCountdown: () => this.processTerrainCountdown(),
+      processStatusDamage: () => this.processStatusDamage(),
+      processScreenCountdown: () => this.processScreenCountdown(),
+      processTailwindCountdown: () => this.processTailwindCountdown(),
+      processTrickRoomCountdown: () => this.processTrickRoomCountdown(),
+      processHeldItemEndOfTurn: () => this.processHeldItemEndOfTurn(),
+      processLeechSeed: () => this.processLeechSeed(),
+      processPerishSong: () => this.processPerishSong(),
+      processCurse: () => this.processCurse(),
+      processNightmare: () => this.processNightmare(),
+      processBindDamage: () => this.processBindDamage(),
+      processSaltCureEoT: () => this.processSaltCureEoT(),
+      processDefrost: () => this.processDefrost(),
+      processSafeguardCountdown: () => this.processSafeguardCountdown(),
+      processMysteryBerry: () => this.processMysteryBerry(),
+      processStatBoostingItems: () => this.processStatBoostingItems(),
+      processHealingItems: () => this.processHealingItems(),
+      processEncoreCountdown: () => this.processEncoreCountdown(),
+      getOpponentActive: (side) => this.getOpponentActive(side),
+      processAbilityResult: (result, pokemon, opponent, sideParam) =>
+        this.processAbilityResult(result, pokemon, opponent, sideParam),
+      processItemResult: (result, pokemon, sideParam) =>
+        this.processItemResult(result, pokemon, sideParam),
+      checkMidTurnFaints: (moveSource) => this.checkMidTurnFaints(moveSource),
+      applyPrimaryStatus: (target, status, side, sleepTurnsOverride) =>
+        this.applyPrimaryStatus(target, status, side, sleepTurnsOverride),
+    });
   }
 
   private processWeatherDamage(): void {
